Move DNS resolver to pre-allocated thread pool
