+++ src/core.py
@@ -22,6 +22,8 @@ MAX_DEFAULT_THREADS = 4000
 
 WORK_STEALING_DISABLED = -1
 
+DNS_WORKERS = 10
+
 class cl:
     MAGENTA = Fore.LIGHTMAGENTA_EX
     BLUE = Fore.LIGHTBLUE_EX
