+++ src/dns_utils.py
@@ -1,4 +1,4 @@
-from concurrent.futures import ThreadPoolExecutor
+from concurrent.futures import Executor
 from functools import lru_cache
 from typing import List, Optional
 
@@ -34,10 +34,9 @@ def safe_resolve_host(host: str) -> Optional[str]:
         return None
 
 
-def resolve_all_targets(targets: List[Target], max_workers:int=32) -> List[Target]:
+def resolve_all_targets(targets: List[Target], pool: Executor) -> List[Target]:
     unresolved_hosts = list(set(target.url.host for target in targets if not target.is_resolved))
-    with ThreadPoolExecutor(max_workers=max_workers) as pool:
-        ips = dict(zip(unresolved_hosts, pool.map(safe_resolve_host, unresolved_hosts)))
+    ips = dict(zip(unresolved_hosts, pool.map(safe_resolve_host, unresolved_hosts)))
     for target in targets:
         if not target.is_resolved:
             target.addr = ips.get(target.url.host)
